phase 3: scope contraction — delete dead handlers, services
Drop all dns-api-go surface area with zero callers across the Spinup +
SpinupManaged ecosystems: /zones, /networks, /macs, /id/{id}, /search,
/systeminfo, and /ips/{ip} GET. Remove their backing services
(BaseService, ZoneService, NetworkService, MacAddressService) and the
interfaces only they implemented. Preserve the 7-route surface
server-api
and SpinupManaged consume, plus the v1 helpers RecordService and
IpAddressService still rely on (those retire in Phase 4 when those
services migrate to v2). Single deletion commit, no v2 logic added.
Net -1,515 lines; offline tests and 5 live tests against BAM-test pass.
