+++ internal/api/routes.go
@@ -28,7 +28,6 @@ func (s *server) routes() {
 	api.HandleFunc("/version", s.VersionHandler).Methods(http.MethodGet)
 	api.Handle("/metrics", promhttp.Handler()).Methods(http.MethodGet)
 	api.HandleFunc("/", s.HomeHandler).Methods(http.MethodGet)
-	api.HandleFunc("/systeminfo", s.SystemInfoHandler).Methods(http.MethodGet)
 
 	// Create a subrouter for routes that need account validation
 	accountRouter := api.PathPrefix("/{account}").Subrouter()
@@ -36,35 +35,14 @@ func (s *server) routes() {
 	// Apply the middleware to all routes in this subrouter
 	accountRouter.Use(s.AccountValidationMiddleware)
 
-	// Custom search based on type and filters
-	accountRouter.HandleFunc("/search", s.CustomSearchHandler).Methods(http.MethodGet)
-
-	// Manage entities by ID
-	accountRouter.HandleFunc("/id/{id}", s.GetEntityHandler()).Methods(http.MethodGet)
-	accountRouter.HandleFunc("/id/{id}", s.DeleteEntityHandler()).Methods(http.MethodDelete)
-
-	// Manage Zones
-	accountRouter.HandleFunc("/zones", s.GetZonesHandler()).Methods(http.MethodGet)
-	accountRouter.HandleFunc("/zones/{id}", s.GetZoneHandler()).Methods(http.MethodGet)
-
 	// Manage DNS records
 	accountRouter.HandleFunc("/records", s.GetRecordsHandler).Methods(http.MethodGet)
 	accountRouter.HandleFunc("/records/{id}", s.GetRecordHandler()).Methods(http.MethodGet)
 	accountRouter.HandleFunc("/records/{id}", s.DeleteRecordHandler()).Methods(http.MethodDelete)
 	accountRouter.HandleFunc("/records", s.CreateRecordHandler).Methods(http.MethodPost)
 
-	// Manage Networks
-	accountRouter.HandleFunc("/networks", s.GetNetworksHandler()).Methods(http.MethodGet)
-	accountRouter.HandleFunc("/networks/{id}", s.GetNetworkHandler()).Methods(http.MethodGet)
-
 	// Manage IP addresses
 	accountRouter.HandleFunc("/ips/cidrs", s.GetCIDRHandler).Methods(http.MethodGet)
-	accountRouter.HandleFunc("/ips/{ip}", s.GetIpAddressHandler).Methods(http.MethodGet)
 	accountRouter.HandleFunc("/ips/{ip}", s.DeleteIpAddressHandler).Methods(http.MethodDelete)
 	accountRouter.HandleFunc("/ips", s.AssignIpAddressHandler).Methods(http.MethodPost)
-
-	// Manage MAC addresses
-	accountRouter.HandleFunc("/macs/{mac}", s.GetMacAddressHandler).Methods(http.MethodGet)
-	accountRouter.HandleFunc("/macs", s.CreateMacAddressHandler).Methods(http.MethodPost)
-	accountRouter.HandleFunc("/macs/{mac}", s.UpdateMacAddressHandler).Methods(http.MethodPut)
 }
