--- internal/interfaces/entities_by_hint_lister.go
@@ -1,7 +1,0 @@
-package interfaces
-
-import "dns-api-go/internal/models"
-
-type EntitiesByHintLister interface {
-	GetEntitiesByHint(start int, count int, options map[string]string) (*[]models.Entity, error)
-}
