--- internal/interfaces/entity_updater.go
@@ -1,7 +1,0 @@
-package interfaces
-
-import "dns-api-go/internal/models"
-
-type EntityUpdater interface {
-	UpdateEntity(entity *models.Entity) error
-}
