--- internal/models/mac.go
@@ -1,7 +1,0 @@
-package models
-
-type Mac struct {
-	Address    string
-	PoolId     int
-	Properties map[string]string
-}
